Load the test files in the Rakefile
