+++ test/composition_new_api.rb
@@ -1,3 +1,5 @@
+require "test_helper"
+
 class FormCompositionInheritanceTest < MiniTest::Spec
   module SizePrice
     include Reform::Form::Module
