+++ test/form_new_api.rb
@@ -1,3 +1,5 @@
+require "test_helper"
+
 class FormTest < MiniTest::Spec
   Artist = Struct.new(:name)
