+++ test/inherit_new_api.rb
@@ -1,3 +1,4 @@
+require "test_helper"
 require "representable/json"
 
 class InheritTest < BaseTest
