+++ test/populate_new_api.rb
@@ -1,3 +1,5 @@
+require "test_helper"
+
 class PopulatorTest < MiniTest::Spec
   Song  = Struct.new(:title, :album, :composer)
   Album = Struct.new(:name, :songs, :artist)
