+++ test/save_old_api.rb
@@ -1,3 +1,5 @@
+require "test_helper"
+
 class SaveTest < BaseTest
   Song  = Struct.new(:title, :album, :composer)
   Album = Struct.new(:name, :songs, :artist)
