+++ test/validate_new_api.rb
@@ -1,3 +1,5 @@
+require "test_helper"
+
 class ContractValidateTest < MiniTest::Spec
   Song  = Struct.new(:title, :album, :composer)
   Album = Struct.new(:name, :songs, :artist)
