+++ test/validation/dry_validation_new_api.rb
@@ -1,5 +1,4 @@
-# encoding: utf-8
-
+require "test_helper"
 require "reform/form/dry"
 require "reform/form/coercion"
 #---
