+++ test/validation/dry_validation_old_api.rb
@@ -1,4 +1,4 @@
-# encoding: utf-8
+require "test_helper"
 require "reform/form/dry"
 require "reform/form/coercion"
